correcting the output of the csv based on using references to other docs
